requirements.txt
absl-py==0.9.0
alabaster==0.7.10
apturl==0.5.2
astor==0.6.2
audioread==2.1.8
Babel==2.6.0
backcall==0.1.0
beautifulsoup4==4.4.1
bleach==1.5.0
blinker==1.3
blurhash==1.1.4
boost==0.1
Brlapi==0.6.4
bsddb3==6.1.0
cachetools==4.0.0
certifi==2018.11.29
cffi==1.13.2
chardet==3.0.4
checkbox-support==0.22
Click==7.0
cloudpickle==0.5.3
cmake==3.15.3
command-not-found==0.3
constants==0.6.0
cryptography==1.2.3
cycler==0.10.0
Cython==0.28.3
dask==0.17.5
decorator==4.3.0
defer==1.0.6
devscripts===2.16.2ubuntu3
dlib==19.19.0
docopt==0.6.2
docutils==0.14
entrypoints==0.2.3
facenet-pytorch==1.0.1
feedparser==5.1.3
flake8==2.5.4
gast==0.2.2
google-auth==1.10.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
graphviz==0.8.4
grpcio==1.11.0
guacamole==0.9.2
h5py==2.7.1
html5lib==0.9999999
httplib2==0.9.1
idna==2.8
imagesize==1.0.0
imgaug==0.2.5
imutils==0.5.3
ipykernel==4.8.2
ipyparallel==6.2.1
ipython==7.9.0
ipython-genutils==0.2.0
ipywidgets==7.2.1
jedi==0.12.0
Jinja2==2.8
joblib==0.14.1
jsonpatch==1.24
jsonpointer==2.0
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.0.1
language-selector==0.1
librosa==0.7.1
llvmlite==0.31.0
louis==2.6.4
lxml==3.5.0
Magic-file-extensions==0.2
Mako==1.0.3
Markdown==2.6.11
MarkupSafe==0.23
mask-rcnn==2.1
Mastodon.py==1.5.0
matplotlib==2.2.2
mccabe==0.2.1
mistune==0.8.3
mtcnn==0.1.0
mxnet==1.3.1
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
nose==1.3.7
notebook==5.5.0
numba==0.47.0
numpy==1.17.4
oauthlib==3.1.0
onboard==1.2.0
opencv-contrib-python==4.1.2.30
openslide-python==1.1.0
opt-einsum==3.1.0
packaging==17.1
padme==1.1.1
pandas==0.23.1
pandocfilters==1.4.2
parso==0.2.0
pep8==1.7.0
pexpect==4.0.1
pickleshare==0.7.4
Pillow==5.1.0
pkg-resources==0.0.0
plainbox==0.25
prompt-toolkit==2.0.10
protobuf==3.11.1
ptyprocess==0.5
pyasn1==0.1.9
pyasn1-modules==0.2.7
pycocotools==2.0.0
pycparser==2.19
pycups==1.9.73
pycurl==7.43.0
pyflakes==1.1.0
Pygments==2.2.0
pygobject==3.20.0
PyJWT==1.3.0
PyOpenGL==3.0.2
pyparsing==2.0.3
python-apt==1.1.0b1+ubuntu0.16.4.5
python-dateutil==2.7.3
python-debian==0.1.27
python-magic==0.4.15
python-speech-features==0.6
python-systemd==231
pytz==2018.4
PyWavelets==0.5.2
pyxdg==0.25
PyYAML==3.12
pyzmq==17.0.0
qtconsole==4.3.1
reportlab==3.3.0
requests==2.21.0
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.0
scenedetect==0.5.1.1
scikit-image==0.14.0
scikit-learn==0.19.1
scipy==1.0.1
screen-resolution-extra==0.0.0
seaborn==0.9.0
Send2Trash==1.5.0
sentencepiece==0.1.85
sessioninstaller==0.0.0
simplegeneric==0.8.1
six==1.10.0
sklearn==0.0
snowballstemmer==1.2.1
SoundFile==0.10.3.post1
Sphinx==1.7.5
sphinxcontrib-websupport==1.0.1
SQLAlchemy==1.3.12
ssh-import-id==5.5
system-service==0.3
tb-nightly==1.9.0a20180605
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.8.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
Theano==1.0.2
toolz==0.9.0
torch==1.1.0
torchfile==0.1.0
torchvision==0.3.0
tornado==5.0.2
tqdm==4.41.1
traitlets==4.3.2
typing==3.6.4
ubuntu-drivers-common==0.0.0
ufw==0.35
Unidecode==1.1.1
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-gdrive==0.7
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urllib3==1.24.1
usb-creator==0.3.0
utils==0.9.0
virtualenv==15.1.0
visdom==0.1.8.9
wcwidth==0.1.7
websocket-client==0.56.0
Werkzeug==0.14.1
widgetsnbextension==3.2.1
wrapt==1.11.2
xdiagnose==3.8.4.1
xkit==0.0.0
XlsxWriter==0.7.3
yamlordereddictloader==0.4.0
